+++ docs/Changelog.md
@@ -3,6 +3,7 @@
 ## v2.3.2
 * Disable auto feed refresh by default and add setting to re-enable it
 * Update theme
+* Disable bugsnag completely when opting out of analytics
 
 ## v2.3.1
 * Hide all story panels if enabled